Base state indicator moved to HUD (domination gamemode)
Having the nearest base state bar rendered right next to its flag (3D) was confusing.
It is now moved to HUD and drawn under the compass for now.
Later, we could and should allow the players to choose exactly where the
bar should be drawn (as well as its width and height?).