Fix server sending bad saved scores
Client expects a sequence of saved scores, terminated by
a saved score with team id -1. Among the fields that it expects
in each saved score is user id, which server does not send.
Consequently, client will interpret some other fields as
user id. If that data includes terminating -1, it will
enter infinite loop, reading score records from past the
packet end and eventually crashing when out of memory.

The solution is for server to send user id, as client expects.